Add files via upload
The Airbnb phenomenon has transformed the way visitors experience New York in the city that never sleeps. In this research, I will take a data-driven tour of the Airbnb listings in New York; each borough has an own tale to tell utilizing PowerBI to reveal the trends, patterns, and insights that create this thriving business.